+++ packages/parcel-config-pwa-manifest/__tests__/main.test.ts
@@ -2,7 +2,6 @@ import Parcel, { createWorkerFarm } from '@parcel/core';
 import { MemoryFS } from '@parcel/fs';
 import { join } from 'path';
 
-
 process.chdir(__dirname);
 
 const fp = (...p: string[]): string => join(__dirname, ...p);
@@ -20,14 +19,26 @@ test('Integrated correctly', async () => {
   const { bundleGraph } = await parcel.run();
   const bundles = bundleGraph.getBundles();
   expect(
-    (await outputFS.readFile(bundles.find(b => b.filePath.endsWith('index.html'))!.filePath, 'utf8'))
-      .replace(/<script(.*)>/, '')
+    (
+      await outputFS.readFile(
+        bundles.find(b => b.filePath.endsWith('index.html'))!.filePath,
+        'utf8'
+      )
+    ).replace(/<script(.*)>/, '')
   ).toMatchSnapshot();
   expect(
-    await outputFS.readFile(bundles.find(b => b.filePath.endsWith('.xml'))!.filePath, 'utf8')
+    await outputFS.readFile(
+      bundles.find(b => b.filePath.endsWith('.xml'))!.filePath,
+      'utf8'
+    )
   ).toMatchSnapshot();
   expect(
-    JSON.parse(await outputFS.readFile(bundles.find(b => b.filePath.endsWith('.webmanifest'))!.filePath, 'utf8'))
+    JSON.parse(
+      await outputFS.readFile(
+        bundles.find(b => b.filePath.endsWith('.webmanifest'))!.filePath,
+        'utf8'
+      )
+    )
   ).toMatchSnapshot();
   await workerFarm.end();
 });